+++ package.json
@@ -14,9 +14,12 @@
     "@tensorflow-models/posenet": "^2.2.2",
     "@tensorflow/tfjs": "^4.16.0",
     "axios": "^1.6.5",
+    "buffer": "^6.0.3",
     "dotenv": "^16.3.2",
     "martinez-polygon-clipping": "^0.7.3",
     "matter-js": "^0.19.0",
+    "mock-socket": "^9.3.1",
+    "path-browserify": "^1.0.1",
     "pixi.js": "^7.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -24,6 +27,7 @@
     "react-redux": "^9.1.0",
     "react-router-dom": "^6.21.3",
     "react-transition-group": "^4.4.5",
+    "rollup-plugin-node-polyfills": "^0.2.1",
     "socket.io-client": "^4.7.4",
     "styled-components": "^6.1.8"
   },
