+++ src/mocks/Browser.ts
@@ -3,3 +3,5 @@ import { setupWorker } from 'msw/browser';
 import { handlers } from './Handlers.ts';
 
 export const worker = setupWorker(...handlers);
+
+worker.start();
