+++ src/types/game.ts
@@ -1,10 +1,10 @@
-import { RoomData } from './room';
+import { RoomData, UserProfile } from './room';
 
 export type GameState = {
     rooms: RoomData[];
     currentPage: number;
     isModalOpen: boolean;
     selectedRoom: RoomData | null;
-    player:RoomData;
+    player:UserProfile;
 }
